fix: relocated metric related components and updated consistency with top buttons.
