+++ commitment/imports/ui/components/metrics-page/DatePickerButton.tsx
@@ -44,7 +44,7 @@ export function DateRangePicker({ onChange, defaultValue }: Props) {
         <Button
           variant="outline"
           className={cn(
-            "w-[280px] justify-start text-left font-normal border-2 rounded-xl",
+            "w-[300px] justify-start text-left font-normal border-2 rounded-lg",
             !date && "text-muted-foreground"
           )}
           style={{ borderColor: "#35353140" }}
