+++ commitment/imports/ui/components/metrics-page/MetricsPage.tsx
@@ -2,11 +2,11 @@ import React, { useState, useMemo } from "react";
 import { format, subDays, addDays } from "date-fns";
 import { DateRange } from "react-day-picker";
 import InfoButton from "../ui/infoButton";
-import { DateRangePicker } from "../ui/datePicker";
-import { BranchDropDownMenu } from "../ui/branchDropDownMenu";
-import UserContributionHeatMap from "./heatmap";
+import { DateRangePicker } from "./DatePickerButton";
+import { BranchDropDownMenu } from "./BranchDropDownMenuButton";
+import UserContributionHeatMap from "./HeatMapGraph";
 import { dark2 } from "../ui/colors";
-import { ContributorDropDownMenu } from "../ui/contributorDropDownMenu";
+import { ContributorDropDownMenu } from "./ContributorDropDownButton";
 import { HighlightCardWithGraph } from "./HighlightCard";
 import { ContributorLineGraph } from "./LineGraph";
 import { LeaderboardGraph } from "./LeaderboardGraph";
